:curly_loop: The project has been updated to target .NET 8.0, with several package references also updated to newer versions. This change is reflected in both `DateOnlyDataProviderApp.csproj` and `DateOnlyApp.csproj` files. The Dapper library has been introduced to the project, with a new class `DateOnlyDataProviderApp.Handlers` and a `DataReaderDapperExample` method that uses Dapper to execute a SQL query and return a list of `VisitorLog` objects. This is demonstrated in `DataOperations.cs` and `Program.cs`.
The `DateOnlyConverter.cs` file and `Context.cs` have been updated with comments indicating that certain classes and methods are not needed when using EF Core 8 and higher. The `ConfigureConventions` method in `Context.cs` has been commented out and replaced with a note that it can be removed when using EF Core 8 and higher.

The `VisitorLog.cs` has seen the removal of the `VisitOn` property from the `VisitorLog` class. In `Person.cs`, a comment has been added to the `BirthDate` property indicating that the `ConfigureConventions` method in the `DbContext` should be referred to for pre EF Core 8 configurations.

The `appsettings.json` file has been simplified by removing the "Stage" and "Production" connection strings. Two new files, `SqlDateOnlyTypeHandler.cs` and `SqlTimeOnlyTypeHandler.cs`, have been added. These files define type handlers for the `DateOnly` and `TimeOnly` types, respectively, for use with the Dapper library.

Changes:

1. Updated `DateOnlyConverter.cs` with a comment about EF Core 8 and higher.
2. Added a sequence named "seq_test" in `Context.cs` and commented out `ConfigureConventions` method.
3. Simplified the code for getting the database connection string in `Program.cs`.
4. Included Dapper library and a new class in `DataOperations.cs`.
5. Updated `DateOnlyDataProviderApp.csproj` and `DateOnlyApp.csproj` to target .NET 8.0.
6. Added a new section of code in `Program.cs` to demonstrate the use of Dapper.
7. Updated `readme.md` with a note about the project update.
8. Removed `VisitOn` property from `VisitorLog` class.
9. Commented out `ConfigureConventions` method in `Context.cs` and used `ConnectionString` method from `ConfigurationHelper` class.
10. Added several methods in `ProviderExtensions.cs` to extend `SqlDataReader` class.
11. Added a comment to `BirthDate` property in `Person.cs`.
12. Simplified `appsettings.json` by removing the "Stage" and "Production" connection strings.
13. Added `SqlDateOnlyTypeHandler.cs` and `SqlTimeOnlyTypeHandler.cs` files.